+++ packages/shared/src/lib/core/layer-2/helpers/index.ts
@@ -1,3 +1,4 @@
+export * from './calculateGasFeeInGlow'
 export * from './encodeAddress'
 export * from './encodeAssetAllowance'
 export * from './encodeSmartContractParameters'
